configured auth on identity and endpoints on auctioncontroller

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -41,12 +42,13 @@ namespace AuctionService.Controllers
             return mapper.Map<AuctionDto>(auction);
         }
 
+        [Authorize]
         [HttpPost]
         public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto createAuctionDto)
         {
             var auction = mapper.Map<Auction>(createAuctionDto);
-            // TODO: Add the current user to the auction as Seller
-            auction.Seller = "test";
+
+            auction.Seller = User.Identity?.Name ?? "Unknown user";
 
             context.Auctions.Add(auction);
 
@@ -62,6 +64,7 @@ namespace AuctionService.Controllers
                 new { Id = auction.Id }, mapper.Map<AuctionDto>(auction));
         }
 
+        [Authorize]
         [HttpPut("{id}")]
         public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
         {
@@ -70,7 +73,7 @@ namespace AuctionService.Controllers
 
             if (auction == null) return NotFound();
 
-            // TODO: check seller == username
+            if (auction.Seller != User.Identity?.Name) return Forbid();
 
             auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
             auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -87,6 +90,7 @@ namespace AuctionService.Controllers
             return BadRequest("Problem saving changes");
         }
 
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteAuction(Guid id)
         {
@@ -94,7 +98,7 @@ namespace AuctionService.Controllers
 
             if (auction == null) return NotFound();
 
-            // TODO: check seller == username
+            if (auction.Seller != User.Identity?.Name) return Forbid();
 
             context.Auctions.Remove(auction);
